Revert "MWPW-140268 [MILO][MEP] Overhaul MEP Action List (#1941)"
This reverts commit d9535d3ba38fa8f31c9eea5f457d68525add5bab.
