+++ libs/blocks/fragment/fragment.js
@@ -59,7 +59,7 @@ function replaceDotMedia(path, doc) {
 }
 
 export default async function init(a) {
-  const { expFragments, decorateArea, mep } = getConfig();
+  const { expFragments, decorateArea } = getConfig();
   let relHref = localizeLink(a.href);
   let inline = false;
 
@@ -70,9 +70,9 @@ export default async function init(a) {
   }
 
   const path = new URL(a.href).pathname;
-  if (expFragments?.[path] && mep) {
-    relHref = mep.handleFragmentCommand(expFragments[path], a);
-    if (!relHref) return;
+  if (expFragments?.[path]) {
+    a.href = expFragments[path];
+    relHref = expFragments[path];
   }
 
   if (isCircularRef(relHref)) {
