--- libs/mep/mepmoreactions/insertScript.js
@@ -1,9 +1,0 @@
-setTimeout(() => {
-  document.querySelector('main > :last-child').insertAdjacentHTML('beforeend', `<div class="text-block">
-    <div class="foreground">
-      <div data-valign="middle">
-        <p>This text was inserted by insertScript.</p>
-      </div>
-    </div>
-  </div>`);
-}, 1000);
