--- libs/mep/mepmoreactions/quote/quote.js
@@ -1,3 +1,0 @@
-export default function init(el) {
-  el.innerHTML = '<p>This quote block\'s text was replaced by experimental block code.</p>';
-}
