--- test/features/personalization/mocks/actions/manifestInsertBefore.json
@@ -1,28 +1,0 @@
-{
-  "total": 5,
-  "offset": 0,
-  "limit": 5,
-  "data": [
-    {
-      "action": "insertBefore",
-      "selector": ".marquee",
-      "page filter (optional)": "",
-      "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/insertbefore",
-      "firefox": "",
-      "android": "",
-      "ios": ""
-    },
-    {
-      "action": "insertBefore",
-      "selector": "/fragments/insertaround",
-      "page filter (optional)": "",
-      "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/insertbeforefragment",
-      "firefox": "",
-      "android": "",
-      "ios": ""
-    }
-  ],
-  ":type": "sheet"
-}
