--- test/features/personalization/mocks/actions/manifestReplacePage.json
@@ -1,15 +1,0 @@
-{
-  "total": 5,
-  "offset": 0,
-  "limit": 5,
-  "data": [
-    {
-      "action": "replacePage",
-      "selector": "",
-      "page filter (optional)": "",
-      "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/replacePage"
-    }
-  ],
-  ":type": "sheet"
-}
