--- test/features/personalization/mocks/fragments/appendToSection.plain.html
@@ -1,1 +1,0 @@
-mock fragment
