--- test/features/personalization/mocks/fragments/insertafter.plain.html
@@ -1,1 +1,0 @@
-mock fragment
