--- test/features/personalization/mocks/fragments/insertafterfragment.plain.html
@@ -1,1 +1,0 @@
-mock fragment
