--- test/features/personalization/mocks/fragments/insertbefore.plain.html
@@ -1,1 +1,0 @@
-mock fragment
