--- test/features/personalization/mocks/fragments/insertbeforefragment.plain.html
@@ -1,1 +1,0 @@
-mock fragment
