--- test/features/personalization/mocks/fragments/milo-replace-content-chrome-howto-h2.plain.html
@@ -1,1 +1,0 @@
-mock fragment
