--- test/features/personalization/mocks/fragments/prependToSection.plain.html
@@ -1,1 +1,0 @@
-mock fragment
