+++ test/features/personalization/mocks/inlineFragReplaced.plain.html
