+++ test/features/personalization/mocks/manifestInsertContentAfter.json
@@ -8,7 +8,7 @@
       "selector": ".marquee",
       "page filter (optional)": "",
       "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/insertafter",
+      "chrome": "/fragments/insertafter",
       "firefox": "",
       "android": "",
       "ios": ""
