+++ test/features/personalization/mocks/manifestRemove.json
