+++ test/features/personalization/mocks/manifestReplace.json
@@ -9,7 +9,7 @@
       "page filter (optional)": "",
       "param-newoffer=123": "",
       "chrome": "",
-      "firefox": "/test/features/personalization/mocks/fragments/milo-replace-content-firefox-accordion",
+      "firefox": "/drafts/rwoblesk/personalization-testing/fragments/milo-replace-content-firefox-accordion",
       "android": "",
       "ios": ""
     },
@@ -18,8 +18,8 @@
       "selector": "#features-of-milo-experimentation-platform",
       "page filter (optional)": "",
       "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/milo-replace-content-chrome-howto-h2",
-"firefox": "",
+      "chrome": "/fragments/milo-replace-content-chrome-howto-h2",
+      "firefox": "",
       "android": "",
       "ios": ""
     }
