+++ test/features/personalization/mocks/manifestReplaceFragment.json
@@ -8,7 +8,7 @@
       "selector": "/fragments/replaceme",
       "page filter (optional)": "",
       "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/fragmentReplaced",
+      "chrome": "/fragments/fragmentreplaced",
       "firefox": "",
       "android": "",
       "ios": ""
@@ -18,7 +18,7 @@
       "selector": "/fragments/inline-replaceme",
       "page filter (optional)": "",
       "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/inlineFragReplaced",
+      "chrome": "/fragments/inline-fragmentreplaced",
       "firefox": "",
       "android": "",
       "ios": ""
