+++ test/features/personalization/mocks/manifestReplacePage.json
@@ -4,11 +4,11 @@
   "limit": 5,
   "data": [
     {
-      "action": "appendToSection",
-      "selector": "section2",
+      "action": "replacePage",
+      "selector": "",
       "page filter (optional)": "",
       "param-newoffer=123": "",
-      "chrome": "/test/features/personalization/mocks/fragments/appendToSection"
+      "chrome": "/test/features/personalization/mocks/replacePage"
     }
   ],
   ":type": "sheet"
