+++ test/features/personalization/mocks/manifestUpdateMetadata.json
