+++ test/features/personalization/mocks/manifestUseBlockCode.json
