+++ test/features/personalization/mocks/manifestUseBlockCode2.json
