+++ test/features/personalization/mocks/personalization.html
@@ -6,7 +6,7 @@
           <div>
             <div data-valign="middle">
               <picture>
-                <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
               </picture>
             </div>
           </div>
@@ -18,7 +18,7 @@
             </div>
             <div data-valign="middle">
               <picture>
-                <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="811" height="207">
+                <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="811" height="207">
               </picture>
             </div>
           </div>
@@ -35,7 +35,7 @@
           <div>
             <div data-valign="middle">
               <picture>
-                <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
               </picture>
             </div>
             <div data-valign="middle">
@@ -48,7 +48,7 @@
           <div>
             <div data-valign="middle">
               <picture>
-                <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
               </picture>
             </div>
             <div data-valign="middle">
@@ -61,7 +61,7 @@
           <div>
             <div data-valign="middle">
               <picture>
-                <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
               </picture>
             </div>
             <div data-valign="middle">
@@ -81,7 +81,7 @@
               <p>This will explain the basic steps on how to use the Milo Experimentation Platform.</p>
               <p>
                 <picture>
-                  <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                  <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
                 </picture>
               </p>
             </div>
@@ -116,10 +116,6 @@
           </div>
         </div>
         <div class="myblock">This block does not exist</div>
-      </div>
-      <div>
-        <p><a href="/fragments/removeme">/fragments/removeme</a></p>
-        <p><a href="/fragments/insertaround">/fragments/insertaround</a></p>
       </div>
     </main>
     <footer></footer>
