+++ test/features/personalization/mocks/postPersonalization.html
@@ -10,7 +10,7 @@
             <div>
               <div data-valign="middle">
                 <picture>
-                  <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                  <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
                 </picture>
               </div>
             </div>
@@ -22,7 +22,7 @@
               </div>
               <div data-valign="middle">
                 <picture>
-                  <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="811" height="207">
+                  <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="811" height="207">
                 </picture>
               </div>
             </div>
@@ -39,7 +39,7 @@
                 <p>This will explain the basic steps on how to use the Milo Experimentation Platform.</p>
                 <p>
                   <picture>
-                    <img loading="lazy" alt="" src="/test/features/personalization/mocks/media_1.jpg" width="1600" height="914">
+                    <img loading="lazy" alt="" src="/test/mocks/media_1.jpg" width="1600" height="914">
                   </picture>
                 </p>
               </div>
