+++ test/features/personalization/replacePage.test.js
@@ -10,9 +10,9 @@ it('replacePage should replace all of the main block', async () => {
   const config = getConfig();
   config.env = { name: 'prod' };
 
-  let manifestJson = await readFile({ path: './mocks/actions/manifestReplacePage.json' });
+  let manifestJson = await readFile({ path: './mocks/manifestReplacePage.json' });
   manifestJson = JSON.parse(manifestJson);
-  const replacePageHtml = await readFile({ path: './mocks/fragments/replacePage.plain.html' });
+  const replacePageHtml = await readFile({ path: './mocks/replacePage.plain.html' });
 
   window.fetch = stub();
   window.fetch.onCall(0).returns(
